Created py-sqs-delete-specific-messages
This file (script) will pull specific messages, they will use the 'Message_ID' attribute of a message upon pulling that message, this is done to identify this message. The way this works in sqs messages handling is,  they will rely on the 'Receipt_Handle' to identify the specific message, hence the messages will be pulled and couples with their 'Receipt_Handle' . This will then be used to delete the message. The args from sys was used to enable this script to work through passing the sqs url and message id through command line.